Refactor exceptions: add UserException, remove ReviewedHellbenderException, rename HellbenderException to GATKException
The important distinction is not between "reviewed" and "unreviewed" exceptions (all
use of exceptions should be reviewed in code review, after all), but between user mistakes
and internal sanity check failures.

To this end, I've ported UserException from the old GATK codebase (preserving only the
most generally useful subclasses -- we can add more later if needed), removed the silly
ReviewedHellbenderException, and renamed HellbenderException to GATKException, which is
what should be thrown for all errors that are not the user's fault.