Disable test suite parallelism
We've been experiencing intermittent, inexplicable failures in the test suite.
In a recent travis build, an assertion in MisencodedBaseQualityReadTransformerUnitTest
failed for no reason at all, then passed on the re-run (with no code changes).

My initial suspect is the new test suite parallelism, which based on my experience
with TestNG is a potentially inexhaustible source of weird, impossible-to-reproduce errors.
Let's turn this off for now and see if these intermittent failures go away. The
extra speed isn't worth it if we can't trust the results of our travis builds!