reduce iterations in tests using generateCigarList
We have a number of tests using an exhaustive cigar generator ReadClipperTestUtils.generateCigarList to generate every permutation of cigar string up to a certain length.
Reduced it the maximum cigar length in tests in  ReadClipperUnitTest, SplitNCigarReadsIntegrationTest, and SplitNCigarReadsUnitTest from 10 -> 9.

This reduces the time these tests take by ~1 order of magnitude and shouldn't reduce the effectiveness of the tests very much.
Also renamed the length variable in some cases to more accurately describe it.