Add an initial stub end-to-end integration test for the ReadsPreProcessingPipeline
Just checks that the input and output bams are equal (which is still the
case for a bam with no duplicates). Once BQSR is hooked up we'll have
to update the expected output for this test.

This is intended as a starting point for the more meaningful tests we
eventually want to have.