Added data-set writing support for HDF5 files.
HDF5Reader is now called HDF5File and can also be used for writing depending on the file openning mode used.
We have added makeDouble, makeDoubleArray, makeDoubleMatrix, makeStringArray operations to write these types of data into the
HDF5 file; we use "make" instead of "write" as these actually may create a data-set (and enclosing groups recursively)
if none present rather than just overwrite the value of an existing data-set. If we need a set of methods to change the value of an existing
dataset, that set will use the "write" prefix.