Don't check contig ordering during sequence dictionary validation
GATK4 uses contig names, rather than contig indices, and so does not
need to be as strict as GATK3 and require contigs to occur in the same
relative order in two sequence dictionaries, or check for lexicographic
ordering in human dictionaries.

This solves several major usability issues:

-2bit references (used by the spark BQSR) typically contain dictionaries
 with very non-standard contig ordering. Since we query contigs by name,
 we are actually compatible with these references and shouldn't blow up.

-Many VCF dictionaries use lexicographic ordering, and GATK4 would blow
 up on these.

With this change, we also no longer require variant dictionaries to share
a common subset of contigs with each other (just with the reference and/or
reads).

Resolves #1176
Resolves #1024