Ensure that all spark tools apply the same read filters as their walker equivalents
Most ReadWalkers apply the WellformedReadFilter, but their spark equivalents
were not doing so. This creates an inherited method GATKSparkTool.makeReadFilter()
that defaults to the WellformedReadFilter get and automatically applied to the
RDD of reads returned from GATKSparkTool.getReads().

Only BaseRecalibratorSpark needed to override this method to apply a custom
read filter in order to match the walker filtering settings.

Resolves #1158