adding scripts/gatk-submit.sh
this is a script which can be used after running gradle installDist to run spark jobs
it can be used identically to ths build/install/bin/gatk script, but has extra features for dealing with spark

running a spark tool and supplying the option --sparkRunner with DIRECT, SUBMIT, or GCS has special behavior
DIRECT will run the tool in the in memory spark runner
SUBMIT along with an appropriate --sparkMaster will run on an accessible spark cluster using spark-submit
arguments to spark-submit may be specified before the arguments to GATK by separating them with a --
GCS will submit jobs to google dataproc using gcloud
common arguments for spark submit will be adapted to match the gcloud formating
this will fail if gcloud isn't installed

if GATK_GCS_STAGING is specified, the jar will be uploaded and cached in the specified bucket for rapid re-use

input files will not be autouploaded to the cloud

--dry-run may be specified before the --, this will only print the commands that will be run instead of actually running them

Adding DataProcArgumentReplace simple tool to convert spark-submit args into gcloud args.
This conversion is not guaranteed to translate all spark command line options to matching gcloud ones.
If you find options that are not translated or are miss-translated please file an issue.

added settings.gradle so the project name doesn't depend on the folder it's placed in