Make HeaderlessSAMRecordCoordinateComparator safer
Instead of calling setHeader() to temporarily give headerless reads
a header and then calling into htsjdk's SAMRecordCoordinateComparator,
adapt the htsjdk code directly to work with headerless reads. This should
be safer (especially in a multithreaded context), as mutating the objects
being compared within a comparator is a violation of contract.