Updating to latest gatk to get bugfixes regarding spark cluster.
fixed log issues

Changes to make spark cluster code work.

Modifications to make Spark cluster not throw NPE.  Tested on spark cluster to make sure it is fast and it seems as if it is really fast.  Also, seems to work in standalone mode.

Fixed issue with binsize

Added unit test for binsize change.

Had to remove ctx.close() (could not use ctx.stop() either) in order for unit tests to pass.