fix for MarkDuplicatesSpark not writing output bam
fixes #1448

MarkDuplicates spark was not writing an output bam when running with a standalone spark instance if the output path was a relative file path.
This fixes the problem by making any relative file paths into absolute file paths.
Added a check to see if no part files can be found so that errors like this will crash in the future instead of silently failing.

renamed mergeBam to mergeHeaderlessBamFiles
exposed mergetInto for testing

added a test to ensure that mergeInto throws if no part files are found
moved existing ReadsSparkSink test files to a more standard location
minor refactorings in ReadsSparkSinkUnitTest