fixing spark-submit not running with local[*]
removing spark.executor.userClassPathFirst=true seems to fix the problems encountered when running using spark-submit and sparkMaster local
this fixes #1315 and should also fix #1386

Removing it doesn't seem to cause any problems when running under yarn on the cluster.
I'm not certain why we added spark.executor.userClassPathFirst=true in the first place.  I assumed it was necessary, like spark.driver.userClassPathFirst=true is.
It seems plausible that it was only added for symmetry and it was never actually needed in the first place.