Update build.gradle to make sure that tests are always rerunable. (#1719)
* Update build.gradle

make sure that tests are always rerunable.

Before this change you can run tests only once before gradle decides that the task is "UP-TO-DATE". I want to be able to rerun tests multiple times (bonus: guess what I'm testing) and so this caching got in the way. With this change, tasks are always rerunnable

* move jacoco to script