disabled pairHMM binary build on PPC64LE; fixed floating comparison issue
PG is on by default and I have a test to prove it (#1722)

collections 4.0 had an exploit https://issues.apache.org/jira/browse/COLLECTIONS-580

sputnik (#1747)

* sputnik

* badge

fix lib intel deflater extraction (#1742)

libIntelDeflater.so wasn't being placed in the right location, causing it to not be used

it now is placed in build/libIntelDeflater.so, where it was intended to be
  fixes #1739

adding IntelDeflaterTest that checks if the IntelDeflater is loaded at runtime
  this will be skipped if the OS is not some linux variant
  it runs in travis which should prevent us from accidentally breaking this in the future
  making test dependsOn extractIntelDeflater
  fixes #1739

improving test logging so that tests run in travis and on the command line will output a totals section
  this will clearly show any skipped tests

adding an explicit PrintReads call to travis.  this checks that gatk-launch will run PrintReads and exit without failing
  it runs quickly and outputs some useful log messages

small cleanup (#1698)

HaplotypeCaller: the engine-level pieces

Included in this PR are the engine-level dependencies of the HaplotypeCaller
from the dr_runnable_haplotypecaller branch. This includes all parts
of that branch that lie outside of the haplotypecaller package itself, and
mostly consists of a new traversal (AssemblyRegionWalker), additions to
the genotyper package, various additions/tweaks to common utilities
and engine-level classes, and comprehensive tests.

The haplotypecaller package itself has been removed in preparation for its
migration to the gatk-protected repository.

Add a tolerance to the range check for the probability in ActivityProfileState

Remove the range check in ActivityProfileState.setIsActiveProb()

Turns out that the BandPassActivityProfile can generate probabilities
that slightly exceed 1.0 when moving probability mass around. We should
eventually cap at 1.0, but first we must evaluate the effect of such a
cap on the HaplotypeCaller

fixed tab/space

fix build and tests to run on PPC64LE

disabled pairHMM binary build on PPC64LE
fixed floating comparison issue

set FTZ always on jniComputeLikelihoods (#1764)

make it run as root (#1760)

determine the number of spark cores to use from env. var.; disable IntelDeflator testing

don't use dataflow serializablefunction

adding gatk-launch to the jar

packaging gatk-launch in our jar so that it's available to downstream projects
some build.gradle refactoring

add IBM to AUTHORS

add IBM

determine the number of spark cores to use from env. var.

disable IntelDeflator testing; fix bug left behind

don't use dataflow serializablefunction

adding gatk-launch to the jar

packaging gatk-launch in our jar so that it's available to downstream projects
some build.gradle refactoring

add IBM to AUTHORS

add IBM

revised determineDefaultSparkMaster() method after review