Convert ReadsDataSource to use Paths, so it can read inputs from the cloud.
As a result, every walker that takes a BAM as input can now read it from
the cloud directly.

Sample use: ./gatk-launch CountReads -I
"gs://.../CEUTrio.HiSeq.WGS.b37.ch20.4m-12m.NA12878.bam"
(...)
Tool returned:
7275701

Note that this CL doesn't add a prefetcher or multithreading, so the
performance is inferior to that of the NIO proof of concept which was
using both.