Performance improvements for stratified/sorted ReadPileup
This PR includes the following improvements for ReadPileup by using a new
class for store PileupElement (PileupElementTracker):

* If sorting by alignment start is necessary, maintain a sorted PileupElementTracker
* If splitting by sample is already done, keep the stratified PileupElementTracker

Because pre-sorting and stratification is implemented in LIBS, this will
improve performance in LocusWalker and other tools using it. Other
methods, as fixing overlaps (with the samtools approach) will benefit for
this improvements.