Improve error message for no-access and disabled-account cases
before: GATK crashes with a stack trace. The stack trace contains the
info the user needs, but not in the most readable of ways.

now, a much more readable message:
A USER ERROR has occurred: Couldn't read file gs://(...). Error was:
401: Anonymous users does not have storage.objects.get access to object (...).

Additional information (including a stack trace) is displayed if the
user specifies --verbosity=DEBUG