Move to a snapshot of gatk-public based on 4.alpha.2-188-g7332d10 with TandemRepeat fix applied
This moves us to a snapshot of gatk-public that takes the version gatk-protected
currently depends on (4.alpha.2-188-g7332d10) and applies @davidbenjamin 's fix
to the TandemRepeat annotation to it.

This is necessary to unblock @davidbenjamin 's work, because the HaplotypeCaller
tests are failing if we update protected to the latest public HEAD, and although we've
fixed some of the issues there are still some unexplained failures in the concordance tests.