Includes a new tool to generated BAM extract with the reads that evidence for a number of SV called variants.
It takes as input the variant call vcf, the interval where the variants of interest are, the bam file with the input reads and the location
of the assembly fastsq.

The output is a new BAM file only with the reads that are part of the assemblies that were use to call those variants. These include the original input alignment
data for each read including those pairs that were totally unmapped.

When the evidence reads have primary and secondary alignment there is a possibility that some alternative alingments that are redudant will be present in the output.