Replace MannWhitneyU with updated version from GATK3
MannWhitneyU was re-written from scratch in 2016 in GATK3,
but these changes never got ported to GATK4. This new version
produces significantly different results from the version
currently in GATK4, resulting in VERY different values for the
RankSumTest annotations in HaplotypeCaller output.

@meganshand informs me that the updated GATK3 version has been
validated in R, and has much better tests than the old version.

This is a straightforward port of that version with minimal changes:

-Merged "MWUnitTest" and "RankSumUnitTest" from GATK3 into a single
 test class MannWhitneyUUnitTest
-Ported MathUtils.binomialCoefficient() and wrote new test for it
-Updated RankSumTest class and tests as appropriate

I've confirmed that with this change, the RankSum annotations produced
by the GATK4 HaplotypeCaller closely match those produced by the GATK3
HaplotypeCaller

Resolves #2604