Re-enable openPrivateFileWithExplicitCredentials test
Also updated code and documentation to indicate how to properly run this
test. Sadly it has to be done manually because I don't know of a
reasonable way to disable default credentials.

Nevertheless it's good that the test is there even for automated runs,
so we can check that loading the explicit credentials does not break
anything.