Adding in GATKTool option to explode when index is out of date.
Adding the errorOnOutOfDateIndex option, which will cause a
UserException to be thrown when an index file is opened that is out of
date with respect to its data file.  This option defaults to false to
preserve baseline behavior.

Split tests apart and added a method to change the last modified time of a file.

Fixes #1683