travis PR builds now use the correct commit in docker
Previously we had an issue where our travis builds would use the wrong
commit for the docker builds in the travis pull-request builds but not for the push
builds.

This caused the tests from master to run and usually pass.  However,
since we are mounting the test data from the correct commit into the
docker, this would result in confusing mismatches where old tests would
try to run on new test data.
fixes #3216

Fixing the problem by using the $TRAVIS_COMMIT environment variable
instead of the TRAVIS_BRANCH.