First version of Funcotator, a GATK Oncotator analog.
Fixes #3283

Includes functionality to create annotations from GENCODE for HG19 and
HG38 for MNPs and INDELs in all regions of the genome.

There are still some INDEL issues, and some INDEL tests are still
commented out.

This is very much a work in progress and should still be considered a
BETA tool.

Added required test files for Funcotator unit and integration tests to
git lfs.