Additional integration tests for AlignmentUtils
Expands testReadAlignedToRef to take both an assembled haplotype and a
reference haplotype as input.

Also adds four unit tests, two testing the conversion of
read-to-haplotype CIGAR strings to read-to-reference CIGAR strings and
two testing the correct alignment of reads to haplotypes.

The CIGAR test fails when the haplotype and the aligned read contain
overlapping deletions relative to the reference.

The first alignment unit test fails when left align does not update the
read position after left alignment of an 8bp indel at the start of the
read. The second alignment unit test fails due to the read start
position not being adjusted for a preceding indel in the haplotype.