Adds supplementary alignment info into fastq output, also additional tags like NM and AS
Changes the mapping= field format a bit:

@read-name mapping=unmapped -> @read-name mapping=*

@read-name mapping=chr1,100,+,100M -> @read-name mapping=chr1,100,+,100M,40,10,125
The last three numbers are MQ, NM and AS. NM and AS will be omitted in undefined.
If AS is defined byt NM is not then NM is set to '.'.

Chimeric alignments are now represented with records separated with ';' as in SAM's SA:Z: tag.

Bases orientation is still determined by the strand of the first mapping entry.