Fixed sample name reordering bug in GenomicsDBImport (#3667)
Sample names were getting scrambled in GenomicsDBImport when using the --batchSize option. This was caused by the fact that the callset.json files had the sample names globally sorted, the actual data was entered in the order the sample names were input, but each batch was sorted within the batch. This produced a mismatch between data and sample name mapping.

Fixed by using sorted maps everywhere sample names are handled in the tool.

Modified our tests to use an out of order sample map file and multiple batches.