create GATKBaseTest so downstream projects can use BaseTest #2125 #3029 (#3475)
Projects depending on GATK have had trouble using BaseTest because it tries to load specific files from the gatk test resources.  If the project didn't include these files, the tests would crash.  This separates out a new GATK only GATKBaseTest which contains all of the file references but leaves the useful utilities in BaseTest for downstream use.

It makes similar changes in other test classes that had related issues.
fixes #2125 
fixes #3029
Introduces #3771 but due to the pain of rebasing the entire test suite repeatedly, this will be addressed in a follow up PR.  

* Factor out a GATKBaseTest for separate test resources from test utilities in BaseTest
* Repackage CommandLineProgramTest due to project-specific paths
* Refactor TargetsToolsTestUtils to allow downstream projects to use with custom references