ReadClipper: when hardclipping results in an empty read, avoid trying to set a negative start position
When hard clipping a read resulted in an empty read, and the read was sufficiently
close to the start of a contig, we could end up trying to set the read to a negative
start position and blowing up. This patch causes us to return an empty read earlier,
preventing us from setting an invalid start.

Resolves the "IllegalArgumentException during clipping: contig must be non-null and not
equal to *, and start must be >= 1" error reported by many users of the GATK4
HaplotypeCaller

Includes an integration test proving that the issue is resolved for the HaplotypeCaller

Resolves #3466