ReadClipper: when reverting soft-clipped bases of a read at the start of a contig, don't explode if you end up with an empty read
We had another edge-case bug in our clipping code: when calling ReadClipper.revertSoftClippedBases()
on a read at the start of a contig (position == 1), we could end up with an empty read if the cigar
was something like "41S59H", since the reverted soft-clipped bases would just get hard-clipped away.
The method did not correctly handle this case, and blew up with an exception.

Added a unit test for ReadClipper that fails before the fix and passes after it, as well as an
integration test for HaplotypeCaller that also fails before the fix and passes after it.

Resolves #3845