Improved haplotype sampling for active regions with too many haplotypes
The current approach uses the N best haplotypes based on their score which in turn is approximation to a likelihood based on transition probabilities between nodes.

The problem with this approach is that sometimes we might be missing plausible variants due to complex sub-graphs downstream.

So imagine that you have a snp with allele freq .33 (alt 33% , ref 66%), follow by may variants that result in many slighlty different sub-haplotypes down the graph (referene coordinates).

If the total number of those possible sub-haplotypes exceeds the maximum allowed (128 or so by default). Then all those with the reference allele for the first variant would intrinsically have a better likelihood to any haplotype that contains the alternative allele for the first variant. More concretelly for any given alt-haplotype the corresponding ref-haplotype would be twice as likely (66% / 333%). Thus despite that around 1/3 of the reads support the alternative it does not make it to the final selection.

The fix: instead of returning the top haplotypes by likelihood we sample them "randomly" based on their relative likelihoods, so that any variant allele with enough probability density (proportional to the sum of the likelihood of all haplotypes that contain the variant) should make it to the final set as long as we sample enough haplotypes this way.

The inconvenience here is the fact that we introduce stochasticity in the selection of haplotypes however these would give a fairer representation of the variantion present in the active region.

Notice that this only affect those active-regions where the total number of haplotypes is beyond the maximum allowed (128 by default) so in practice that is relative small number. However this problem has been reported more than once where obvious variants don't even make it to the final haplotype list because of this.

You can see this as an attempt to change from MLE based list of haplotypes to a MAP list although this is all a bit of a approximation to that as we actually dont calculate haplotype likelihoods rigurosly but we rather approximate them using read-threading accounts accross the edge of the graph which should approximate the number of reads that support the transition between vertices in the graph.