Extraced a FuncotatorEngine from Funcotator.java
A large refactoring of the architecture of Funcotator to make the
creation of annotations separate from the command line interface.

This will enable better testing of the annotation functionality by
allowing for unit tests that can test the created funcotations directly
(as opposed to requiring integration tests for the same purpose).