Enable Requester Pays support (#5140)
* Enable Requester Pays support

Google Cloud Storage has "requester pays" buckets. When reading from
those buckets, the requester is billed for the bandwidth (normally, it's
the bucket owner who is billed).

This pull request enables this feature with GATK. By default it is
turned off (so there are no unexpected charges). To turn it on, use the
command line argument: "--gcs-project-for-requester-pays"

Example:

$ ./gatk PrintReads --input $INPUT --output /tmp/reads.bam
fails with: com.google.cloud.storage.StorageException: Bucket is requester pays bucket but no user project provided.

$ ./gatk PrintReads --input $INPUT --output /tmp/reads.bam --gcs-project-for-requester-pays $PROJECT
works

This PR also removes the argumentless version of
setGlobalNIODefaultOptions() because it was confusing (it uses the
default values, not those indicated by the user - usually we'd expect
the user's values to be taken into account).