Fix some AlignmentUtils issues (#3494)
This pull request fixes three issues in the AlignmentUtils class:

* The treatment of D-over-D in function applyCigarToCigar() is backward.
* In function createReadAlignedToRef() the read start position passed to the leftAlignIndel() call may be incorrect if the haplotype has an indel relative to reference.
* And when the leftAlignIndel() call drops any leading D operator in the result cigar, the read start position needs to be adjusted accordingly.

It also adds some unit tests to demonstrate the issues:

Expands testReadAlignedToRef to take both an assembled haplotype and a
reference haplotype as input.

Also adds four unit tests, two testing the conversion of
read-to-haplotype CIGAR strings to read-to-reference CIGAR strings and
two testing the correct alignment of reads to haplotypes.

The CIGAR test fails when the haplotype and the aligned read contain
overlapping deletions relative to the reference.

The first alignment unit test fails when left align does not update the
read position after left alignment of an 8bp indel at the start of the
read. The second alignment unit test fails due to the read start
position not being adjusted for a preceding indel in the haplotype.