Added in cDNA strings for intronic variants.
Now intronic variants produce a cDNA string.  This string is created in
a similar way to how the Codon Change String is created for intronic
variants:

If the transcript containing the variant has no exons, the cDNA string
will be empty.

If the transcript in which the variant occurs contains at least 1 exon, the string will be non-empty and of
the form:
    c.e[EXON NUMBER][+|-][BASES FROM EXON][REF ALLELE]>[ALT ALLELE]
Concretely:
    c.e2-1A>G
Where:
     2 = the number of the exon to which the given variant start is closest
    -1 = number of bases away from the exon (1 before)
     A = Reference allele
     G = Alternate allele

Fixes #5187