Now DE_NOVO_START_IN_FRAME and DE_NOVO_START_OUT_FRAME are correct.
Fixed an issue with de novo starts in the 5' UTR.
Before, the de novo start itself would be detected just fine, however
the position in the UTR was not correctly calculated (leading to an
    incorrect and inconsistent calling of in- vs out-of-frame).
This was due to the code assuming that there was only one 5'UTR.  There
is no limit on the number of 5'UTRs a transcript may have.

This is now accounted for and the calculations match with Oncotator's
assement for in- vs out-of-frame for de novo starts.

Fixes #5333