Funcotator now creates default annotations for difficult variants.
Now Funcotator will create a default annotation in the event that it is
not able to annotate a particular allele pair (i.e. when a
    FuncotatorUtils.TranscriptCodingSequenceException is thrown inside
    the call to
    `createGencodeFuncotationOnTranscript`
    within
    `GencodeFuncotationFactory::createFuncotationsHelper`).
This default annotation will be labeled with the VariantClassification
`COULD_NOT_DETERMINE`.

These will typically only appear in the `otherTranscripts` field because
they have the lowest priority of all VariantClassifications, but they
are not suppressed at all (unlike `IGR` variants).  This is by design.

Fixes #5366