Now correctly populates transcript position.
Before the transcript position would always only be the start position
of the variant in the transcript.

The transcript position field should contain either a single number (for
    SNPS) or two numbers separated by an underscore representing the
transcript start and stop positions of the variant.

This fix correctly changes the field to include both start and stop
positions when start != stop.

Fixes #5375