Illustration of how to achieve deterministic positional downsampling.
The random seed is set every time the genome position changes, so that
downsampling is deterministic and, crucially, identical if the reads
are processed in a single iteration, or in parallel (which is the case
with Spark).

PositionalDownsamplerUnitTest passes with these changes even though
Utils.resetRandomGenerator() is no longer called.

Note that this change is for illustration only: it is not suitable for
merging since ReservoirDownsampler shouldn't care about genome
positions, only PositionalDownsampler should.