Fix VCFFuncotation output order bug Fixes https://github.com/broadinstitute/gatk/issues/6173
I added a single line test that shows the problem, and regenerated the other test files.  I didn't write a specific unit test that proves it's solved.  Feel free to do so if you think it's necessary.
I validated the output by adding the field name to the output value and checking it by eye against the header lines.  This could fairly simply made into a unit test if desired.

I'm not sure if there are other large files that need to be regenerated.

I had initially said this bug only affects vcf but I think it happens to Maf output as well.