DRAGEN-GATK HaplotypeCaller updates
DRAGstr first working approximation to matlab scripts

fist go at rebasing everything correctly, now have to deal with the buggy refactor code

Revert "Rewrote leftAlignIndels code -- it now always works, even for multiple indels (#6427)"

This reverts commit 8612288881890345d949b40ead5796183c52c08d. Which brakes the softclip aware cigar building code and has a horrible bug.

bringing modernity to this branch

added a consistency test for DRAGEN mode

a bunch of code cleanup

more cleanup and tests?

Added the matlab files as test

Strengthened trust in SoftclipBases Behavior for overlap base quality adjustments

Fixing a crash in alleleliklehoods and adjusting bq annotation after the quality for overlapping bases change

minor fix and expanded debug output

removing excess bases before hmm scoring seems to improve the resutlts at complex sites... the FRD issue is still relevant...

Fixed genotype posterior calculator to not honor the unphase permutation number factor

more bugfixes to read edge cases at sites with many low quality ends and low mapping quality combined

including reads that only overlap in hardclips (despite the obvious concerns with that)

fixing an unexpected error

flipping

hardclippingsoftclippedBases

creating a frankenbranch with a very intersting change involving realignment of reads and apparent DRAGEN behavior

fixing a bug in 'getAllVariationEvents()' code

flipping the values so it actually works

Changed BQD feather end calculation to account for indels in the read relative to the reference

reverted the non-realignment code

fixing the horrible regression I just wrought on the BQD calculations...

disabling bad debug squack

Parameterizing the option to use original base qualities

threading the truly original alignment through for BQD/FRD to match the bad dragen behavior we are now aware of

cleaning up the branch

if you like me then you should have put a pin in it

responding to reveiw comments, (still havent fixed the debug streams)

deleting a test that was for local files

fixing a silly bug

fixing the logic yet one more time

fixing an issue with the disable-spanning-event-genotyping code that was causing false positivles (or perhaps reasonable behavior)

jumping the gun on takutos fix because it was causing me trouble for this test

refactored the debug output to be something very simple that should/could/maybe is better handled by making a debug level in the debugger. Anyway this is fine for now and probably could be expanded to handle the other debug streams in HC

adding a scary error message to deter people from using contamination downsampling with BQD/FRD (it should be an exception but somebody would complain and it really isn't a sound thing to do unless the disqualified reads are included but they can't be inherently because they don't have scores)

partial val review response work checkpoint
:

responded to most comments for my code in val PR review

Fixed a couple of rare bugs in DRAGEN port code.
Now we can reads dragstr-params values out of gcloud storage directly.

taking the changes from valentin with the removed OQ tag

pulling out the --use-original-alignments-for-genotyping-overlap as if it were so many weeds in my garden.

Revision changes thus far

Fix issues with how we assign STR period and repeat lenghts to positions on a read.
Added more robust testing using Illumina's Matlab script to generate use cases.

Impose a maximum of 40 for the GOP in DRAGstr pairhmm as seem to be the case in DRAGEN.

Changes to meet recommendation except for the param estimation part that is on a separate branch

reverting a mistake where i replaced every every instance of the word 'exists' with 'isEnabled'

Unifying changes with the refactoring of the ReadsDataSource class

A few updates on the revision:

* Solved a couple of self-evaluation javadoc PR comments.
* Added a integration test for ComposeSTRTableFile based on the DRAGEN output on some of the mini references under src/test/resources/large
* Fix an issue that made ComposeSTRTableFile quite slow with reference with a very large number of contigs (eg. funcotator gencode trascripts ~100_000).
* Found and fixed a bug introduced during refactoring making EstimateDragstrParameters produce "garbage" if run in parallel.
* Silience a tedious debug message in SAMReaderQueryingIterator that comes out too often when the traversal intervals are just a few. Now it requires
          like a 1000+ intervals to produce that debug, might be still too low but it works in practice.

Solves james'  comment on DragstrParam toString deficiency

fixed a few dangling issues with the rebase

added the argument to end all arguments (and also completely remove the users ability to interact with our toolkit as they please)

Fixes bug reported by Cwhelan

Addresses some of James' comments and fixes a bug in a unit test for DragstrLocus class

Fixes Dragstr Read Str analyzer failing unit test

allowing nulls in the genotyper because its actually okay